Added validation code so that all of the inputs on the RequirementsEntry grid are checked. Disabled editing on the column 'Baseline' on the RequirementsEntry grid.
Added new pattern Positive or Negative Number in the InputType for validation.
Added override methods for the InputValidator class that takes in inputs of type doubles or ints and makes sure that the pattern you are passing in to validate the input is the same type as the input type.
Eg: I send in a Double and am trying to validate that it is AlphaNumeric type. This will throw an exception.
But if I send in a Double and am trying to match it to a pattern that is a Double, it passes through.

Changed some brackets while I was reading through code out of ocd habits.